try getting tests to run
